roadmap_gen_alpha_v0.4/requirements.txt
fastapi
uvicorn[standard]
requests
pandas
sentence-transformers
faiss-cpu
scikit-learn
matplotlib
pathlib
python-multipart
jinja2
rapidfuzz